Rename registeredServers to objectExplorer
