+++ src/sql/parts/jobManagement/views/jobHistoryTree.ts
@@ -24,7 +24,7 @@ import { IAction } from 'vs/base/common/actions';
 import { IInstantiationService } from 'vs/platform/instantiation/common/instantiation';
 import { generateUuid } from 'vs/base/common/uuid';
 import * as DOM from 'vs/base/browser/dom';
-import { OEAction } from 'sql/parts/registeredServer/viewlet/objectExplorerActions';
+import { OEAction } from 'sql/parts/objectExplorer/viewlet/objectExplorerActions';
 import { Builder, $, withElementById } from 'vs/base/browser/builder';
 import { AgentJobHistoryInfo } from 'sqlops';
 import { Agent } from 'vs/base/node/request';
