+++ src/sql/parts/objectExplorer/common/nodeType.ts
