+++ src/sql/parts/objectExplorer/serverGroupDialog/media/serverGroupDialog.css
