+++ src/sql/parts/objectExplorer/serverGroupDialog/serverGroup.contribution.ts
