+++ src/sql/parts/objectExplorer/viewlet/media/connectionViewlet.css
