+++ src/sql/parts/objectExplorer/viewlet/media/serverTreeActions.css
