+++ src/sql/parts/objectExplorer/viewlet/recentConnectionDataSource.ts
