+++ src/sql/workbench/common/taskUtilities.ts
@@ -24,7 +24,7 @@ import * as sqlops from 'sqlops';
 import nls = require('vs/nls');
 import os = require('os');
 import path = require('path');
-import { IObjectExplorerService } from 'sql/parts/registeredServer/common/objectExplorerService';
+import { IObjectExplorerService } from 'sql/parts/objectExplorer/common/objectExplorerService';
 import { IWorkbenchEditorService } from 'vs/workbench/services/editor/common/editorService';
 import { QueryInput } from 'sql/parts/query/common/queryInput';
 import { DashboardInput } from 'sql/parts/dashboard/dashboardInput';
