+++ src/sqltest/parts/registeredServer/viewlet/serverTreeView.test.ts
@@ -6,7 +6,7 @@
 'use strict';
 
 import { IConnectionProfile } from 'sql/parts/connection/common/interfaces';
-import { ServerTreeView } from 'sql/parts/registeredServer/viewlet/serverTreeView';
+import { ServerTreeView } from 'sql/parts/objectExplorer/viewlet/serverTreeView';
 import { ConnectionManagementService } from 'sql/parts/connection/common/connectionManagementService';
 
 import { Tree } from 'vs/base/parts/tree/browser/treeImpl';
